Valores de referencia mudados na Integral n-1 : quando n=20, integral é 0.0 ; mudancas na main para imprimir o metodo 1 de 0 a 7 e o metodo 2 de 20 a 7

+++ trabalho_integral/integral.h
@@ -1,7 +1,7 @@
 #ifndef INTEGRAL_H
 #define INTEGRAL_H
 
-double integral_n(int n);
-double integral_n_1(int n_1);
+float integral_n(int n);
+float integral_n_1(int n_1);
 
 #endif
